feat: validator main 패키지 이동 및 lint
